requirements.txt
# GUI 依赖
PyQt5>=5.15.0

# 加密库依赖
pycryptodome>=3.15.0
